move components to app directory
