+++ src/components/app/Note/ActionPanel.tsx
@@ -1,5 +1,5 @@
 import React from "react"
-import { TNote } from "../../reducers/notes/types"
+import { TNote } from "../../../reducers/notes/types"
 
 import { Button } from "rebass"
 import { navigate } from "@reach/router"
