+++ src/components/app/Note/__tests__/action-panel.spec.tsx
